+++ Common/settings.manifest
