+++ Texassemble/Texassemble_Desktop_2019.vcxproj.filters
@@ -16,8 +16,11 @@
     </ResourceCompile>
   </ItemGroup>
   <ItemGroup>
-    <Manifest Include="settings.manifest">
+    <Manifest Include="..\Common\settings.manifest">
       <Filter>Resource Files</Filter>
     </Manifest>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\Common\CmdLineHelpers.h" />
   </ItemGroup>
 </Project>
