+++ Texdiag/texdiag.rc
@@ -28,7 +28,7 @@ LANGUAGE LANG_ENGLISH, SUBLANG_ENGLISH_US
 
 // Icon with lowest ID value placed first to ensure application icon
 // remains consistent on all systems.
-IDI_MAIN_ICON           ICON                    "directx.ico"
+IDI_MAIN_ICON           ICON                    "..\Common\directx.ico"
 
 
 #ifdef APSTUDIO_INVOKED
